passage int -> Long ( attribut code)

+++ ProjetSopraFinal/src/main/java/com/example/demo/repository/MaterielRepository.java
@@ -7,7 +7,7 @@ import com.example.demo.model.Materiel;
 
 public interface MaterielRepository extends JpaRepository<Materiel, Long> {
 
-	public Materiel findByCode(@Param("code") Integer code);
+	public Materiel findByCode(@Param("code") Long code);
 
 	public Materiel findByCout(@Param("cout") Integer cout);
 }
